Added SQLFlowTestBase + single data stream test
